Fixing benchmark run for #175 (#182)
* compile bm_saces class without optimizations

* changed names and moved functions

* improved benchmarks definitions
