+++ tests/benchmark/bm_spaces.cpp
@@ -6,47 +6,16 @@
 #include "VecSim/spaces/space_interface.h"
 #include "VecSim/spaces/space_aux.h"
 
-#include "VecSim/spaces/L2/L2.h"
-#include "VecSim/spaces/IP/IP.h"
-
-class BM_VecSimSpaces : public benchmark::Fixture {
-protected:
-    std::mt19937 rng;
-    size_t dim;
-    float *v1, *v2;
-    Arch_Optimization opt;
-
-    BM_VecSimSpaces() {
-        rng.seed(47);
-        opt = getArchitectureOptimization();
-    }
-
-public:
-    void SetUp(const ::benchmark::State &state) {
-        dim = state.range(0);
-        v1 = new float[dim];
-        v2 = new float[dim];
-        std::uniform_real_distribution<double> distrib(-1.0, 1.0);
-        for (size_t i = 0; i < dim; i++) {
-            v1[i] = (float)distrib(rng);
-            v2[i] = (float)distrib(rng);
-        }
-    }
-
-    void TearDown(const ::benchmark::State &state) {
-        delete v1;
-        delete v2;
-    }
-
-    ~BM_VecSimSpaces() {}
-};
+#include "bm_spaces_class.h"
 
 // Defining the generic benchmark flow: if there is support for the optimization, benchmark the
 // function.
 #define BENCHMARK_DISTANCE_F(arch, settings, func)                                                 \
     BENCHMARK_DEFINE_F(BM_VecSimSpaces, arch##_##settings)(benchmark::State & st) {                \
-        if (opt < ARCH_OPT_##arch)                                                                 \
+        if (opt < ARCH_OPT_##arch) {                                                               \
+            st.SkipWithError("This benchmark requires " #arch ", which is not available");         \
             return;                                                                                \
+        }                                                                                          \
         for (auto _ : st) {                                                                        \
             func(v1, v2, &dim);                                                                    \
         }                                                                                          \
@@ -101,6 +70,9 @@ BENCHMARK_DISTANCE_F(SSE, IP_4_Residuals, InnerProductSIMD4ExtResiduals_SSE)
 #endif // SSE
 
 // Naive algorithms
+
+#include "VecSim/spaces/L2/L2.h"
+#include "VecSim/spaces/IP/IP.h"
 
 BENCHMARK_DEFINE_F(BM_VecSimSpaces, NAIVE_IP)(benchmark::State &st) {
     for (auto _ : st) {
